ai_coach/requirements.txt
streamlit
openai
python-dotenv
pdfkit
PyYAML
streamlit-authenticator
streamlit==1.33.0
streamlit-authenticator==0.3.0
PyYAML
bcrypt
